Inline `IfaceConfig` into `Device`
