+++ rust/connlib/tunnel/src/device_channel/device_channel_win.rs
@@ -9,8 +9,6 @@ use std::task::{Context, Poll};
 
 pub(crate) struct DeviceIo;
 
-pub(crate) struct IfaceConfig;
-
 impl DeviceIo {
     pub fn poll_read(&self, _: &mut [u8], _: &mut Context<'_>) -> Poll<std::io::Result<usize>> {
         // Incoming packets will never appear
@@ -26,35 +24,10 @@ impl DeviceIo {
     }
 }
 
-const BOGUS_MTU: usize = 1_500;
-
-impl IfaceConfig {
-    pub(crate) fn mtu(&self) -> usize {
-        BOGUS_MTU
-    }
-
-    pub(crate) fn refresh_mtu(&self) -> Result<usize> {
-        Ok(BOGUS_MTU)
-    }
-
-    pub(crate) async fn add_route(
-        &self,
-        _: IpNetwork,
-        _: &impl Callbacks,
-    ) -> Result<Option<Device>> {
-        let io = DeviceIo {};
-        let config = IfaceConfig {};
-        Ok(Some(Device { io, config }))
-    }
-}
-
 pub(super) async fn create_iface(
     _: &Interface,
     _: &impl Callbacks,
     _: DnsFallbackStrategy,
 ) -> Result<Device> {
-    Ok(Device {
-        config: IfaceConfig {},
-        io: DeviceIo {},
-    })
+    todo!()
 }
